// src/app/models/todo.ts
export interface Todo {
  todoId?: string;       // Optional because it's usually generated by the backend
  uid: string;         // UID of the user
  description: string;          // The task description
  createdAt?: string;    // Optional because it might be set by the backend
  title?: string;        // Optional, as it might not always be used
  completed: boolean;       // Additional status of the task, similar to 'completed'
  dueDate?: string;      // Due date of the task
  status?: string;       // Status of the task (e.g., 'pending', 'completed')
  category?: string;     // Area or category of the task
  assignedBy: string;
  updatedAt?: string;
  targetValue?: any;
  currentValue?: any;
  targetLabel?: string;
  measuredBy?: string;
}
export interface TodoResponse {
  message: string;
  todos: Todo[];
}